requirements.txt
livekit-agents>=0.10.2
livekit-plugins-openai>=0.10.4
livekit-plugins-deepgram>=0.6.8
livekit-plugins-silero>=0.7.2
python-dotenv~=1.0